+++ .github/workflows/rust.yml
@@ -17,7 +17,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Clippy
-      run: cargo clippy -- -Dwarnings
+      run: cargo clippy
     - name: Build
       run: cargo build --verbose
     - name: Run tests
